# Conflicts:
#	.gitlab-ci.yml
#	README.md
#	arkos/__init__.py
#	arkos/applications.py
#	arkos/backup.py
#	arkos/certificates.py
#	arkos/config.py
#	arkos/connections.py
#	arkos/daemon.py
#	arkos/databases.py
#	arkos/languages/nodejs.py
#	arkos/languages/php.py
#	arkos/languages/python.py
#	arkos/languages/ruby.py
#	arkos/messages.py
#	arkos/security.py
#	arkos/shared_files.py
#	arkos/sharers.py
#	arkos/signals.py
#	arkos/storage.py
#	arkos/system/domains.py
#	arkos/system/filesystems/filesystems.py
#	arkos/system/filesystems/losetup.py
#	arkos/system/groups.py
#	arkos/system/network.py
#	arkos/system/services.py
#	arkos/system/stats.py
#	arkos/system/sysconfig.py
#	arkos/system/systemtime.py
#	arkos/system/users.py
#	arkos/tracked_services.py
#	arkos/updates.py
#	arkos/utilities/__init__.py
#	arkos/utilities/detect.py
#	arkos/utilities/errors.py
#	arkos/utilities/logs.py
#	arkos/utilities/utils.py
#	arkos/websites.py
#	requirements.txt
#	setup.py

+++ arkos/system/filesystems/filesystems.py
@@ -16,6 +16,7 @@ from . import crypto
 from . import losetup
 
 from arkos import config, signals, sharers
+from arkos.messages import Notification, NotificationThread
 from arkos.utilities import shell
 
 libc = ctypes.CDLL(ctypes.util.find_library("libc"), use_errno=True)
@@ -157,6 +158,7 @@ class DiskPartition:
 
 
 class VirtualDisk:
+    """Class to manage virtual disk image objects."""
 
     def __init__(
             self, id_="", path="", mountpoint=None, size=0, fstype="ext4",
@@ -180,37 +182,51 @@ class VirtualDisk:
         self.enabled = enabled
         self.crypt = crypt
 
-    def create(self, mount=False):
+    def create(self, mount=False, will_crypt=False,
+               nthread=NotificationThread()):
         """
         Create virtual disk image.
 
         :param bool mount: Mount after creation?
+        :param bool will_crypt: Will this disk be encrypted later?
+        :param NotificationThread nthread: notification thread to use
         """
+        nthread.title = "Creating virtual disk"
+
         vdisk_dir = config.get("filesystems", "vdisk_dir")
-        if not os.path.exists(os.path.join
-                              (config.get("filesystems", "vdisk_dir"))):
-            os.mkdir(os.path.join(config.get("filesystems", "vdisk_dir")))
+        if not os.path.exists(vdisk_dir):
+            os.mkdir(vdisk_dir)
         self.path = str(os.path.join(vdisk_dir, self.id+".img"))
         if os.path.exists(self.path):
             raise Exception("This virtual disk already exists")
-        signals.emit("filesystems", "pre_add", self)
+
         # Create an empty file matching disk size
+        signals.emit("filesystems", "pre_add", self)
+        msg = "Creating virtual disk..."
+        nthread.update(Notification("info", "Filesystems", msg))
         with open(self.path, "wb") as f:
             written = 0
             with open("/dev/zero", "r") as zero:
                 while self.size > written:
                     written += 1024
                     f.write(zero.read(1024))
-        # Get a free loopback device and mount
-        loop = losetup.find_unused_loop_device()
-        loop.mount(str(self.path), offset=1048576)
-        # Make a filesystem
-        s = shell("mkfs.ext4 {0}".format(loop.device))
-        if s["code"] != 0:
-            excmsg = "Failed to format loop device: {0}"
-            raise Exception(excmsg.format(s["stderr"]))
-        loop.unmount()
-        signals.emit("filesystems", "pre_add", self)
+
+        if not will_crypt:
+            # Get a free loopback device and mount
+            loop = losetup.find_unused_loop_device()
+            loop.mount(str(self.path), offset=1048576)
+            # Make a filesystem
+            msg = "Writing filesystem..."
+            nthread.update(Notification("info", "Filesystems", msg))
+            s = shell("mkfs.ext4 {0}".format(loop.device))
+            if s["code"] != 0:
+                excmsg = "Failed to format loop device: {0}"
+                raise Exception(excmsg.format(s["stderr"]))
+            loop.unmount()
+            msg = "Virtual disk created successfully"
+            nthread.complete(Notification("success", "Filesystems", msg))
+
+        signals.emit("filesystems", "post_add", self)
         if mount:
             self.mount()
 
@@ -225,8 +241,7 @@ class VirtualDisk:
         signals.emit("filesystems", "pre_mount", self)
         if not os.path.isdir(os.path.join("/media", self.id)):
             os.makedirs(os.path.join("/media", self.id))
-        mount_point = self.mountpoint if self.mountpoint\
-            else os.path.join("/media", self.id)
+        mount_point = self.mountpoint or os.path.join("/media", self.id)
         luks_point = os.path.join("/dev/mapper", self.id)
         # Find a free loopback device and mount
         loop = losetup.find_unused_loop_device()
@@ -240,7 +255,7 @@ class VirtualDisk:
                 raise Exception(excmsg.format(self.id, str(s)))
             s = libc.mount(ctypes.c_char_p(bytes(luks_point, 'utf-8')),
                            ctypes.c_char_p(bytes(mount_point, 'utf-8')),
-                           ctypes.c_char_p(bytes(self.fstype)), 0,
+                           ctypes.c_char_p(bytes(self.fstype, 'utf-8')), 0,
                            ctypes.c_char_p(b""))
             if s == -1:
                 crypto.luks_close(self.id)
@@ -253,9 +268,9 @@ class VirtualDisk:
             raise Exception(excstr)
         else:
             s = libc.mount(ctypes.c_char_p(loop.device),
-                           ctypes.c_char_p(mount_point),
-                           ctypes.c_char_p(self.fstype), 0,
-                           ctypes.c_char_p(""))
+                           ctypes.c_char_p(bytes(mount_point, 'utf-8')),
+                           ctypes.c_char_p(bytes(self.fstype, 'utf-8')), 0,
+                           ctypes.c_char_p(b""))
             if s == -1:
                 loop.unmount()
                 excstr = "Failed to mount {0}: {1}"
@@ -308,7 +323,8 @@ class VirtualDisk:
                 self.enabled = False
                 break
 
-    def encrypt(self, passwd, cipher="", keysize=0, mount=False):
+    def encrypt(self, passwd, cipher=config.get("filesystems", "cipher"),
+                keysize=config.get("filesystems", "keysize"), mount=False):
         """
         Encrypt virtual disk image.
 
@@ -589,9 +605,10 @@ def save_fstab_entry(e, remove=False):
 
 def get_partition_uuid_by_name(p):
     """Get a partition's UUID from its device name."""
-    return shell("blkid -o value -s UUID " + p)["stdout"].split("\n")[0]
+    return shell("blkid -o value -s UUID " + p)["stdout"].split(b"\n")[0]
+
 
 
 def get_partition_name_by_uuid(u):
     """Get a partition's device name from its UUID."""
-    return shell("blkid -U " + u)["stdout"].split("\n")[0]
+    return shell("blkid -U " + u)["stdout"].split(b"\n")[0]
